requirements.txt
Django==3.0.3
gunicorn==20.0.4
djangorestframework==3.11.0
markdown==3.1.1
django-filter==2.2.0
django-cors-headers==3.2.1
django-rest-auth==0.9.5
certifi==2020.4.5.1
chardet==3.0.4 
defusedxml==0.6.0 
django-allauth==0.41.0 
idna==2.9 
oauthlib==3.1.0 
python3-openid==3.1.0 
requests==2.23.0 
requests-oauthlib==1.3.0 
urllib3==1.25.8
mysqlclient==1.4.6
django-braces==1.14.0 
django-model-utils==4.0.0
django-private-chat==0.2.2
websockets==3.2